[Silver III] Title: 주유소, Time: 388 ms, Memory: 34944 KB -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -8,9 +8,9 @@ public class Main {
     static StringTokenizer st;
     static StringBuilder sb = new StringBuilder();
     static int N;
-    static int[][] P;
-    static int[] D;
-    static int[] DSum;
+    static long[][] P;
+    static long[] D;
+    static long[] DSum;
 
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
@@ -18,16 +18,16 @@ public class Main {
 
         int N = Integer.parseInt(br.readLine());
 
-        D = new int[N - 1];
+        D = new long[N - 1];
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < N - 1; i++)
             D[i] = Integer.parseInt(st.nextToken());
 
-        DSum = new int[N];
+        DSum = new long[N];
         for (int i = N - 2; i >= 0; i--)
             DSum[i] = DSum[i + 1] + D[i];
 
-        P = new int[N][2];
+        P = new long[N][2];
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < N; i++) {
             P[i][0] = i;
@@ -36,16 +36,16 @@ public class Main {
 
         Arrays.sort(P, (o1, o2) -> {
             if (o1[1] == o2[1])
-                return o1[0] - o2[0];
-            return o1[1] - o2[1];
+                return Math.toIntExact(o1[0] - o2[0]);
+            return Math.toIntExact(o1[1] - o2[1]);
         });
 
         int total = 0;
-        int l = N - 1;
-        for (int[] p : P) {
+        long l = N - 1;
+        for (long[] p : P) {
             if (p[0] == N - 1 || p[0] >= l)
                 continue;
-            total += p[1] * (DSum[p[0]] - DSum[l]);
+            total += p[1] * (DSum[Math.toIntExact(p[0])] - DSum[Math.toIntExact(l)]);
             l = p[0];
         }
